#食品检测项目和用户审核

+++ src/main/java/com/food/sas/service/IAdministratorService.java
@@ -14,5 +14,5 @@ public interface IAdministratorService {
 
     boolean verifyUser(Integer id);
 
-    UserVerificationDTO getUserVerification(Integer i);
+    UserVerificationDTO getUserVerification(Long i);
 }
